pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

# ---------------------------------------------------------------------------
# Project metadata
# ---------------------------------------------------------------------------
[project]
name = "pulseboard-rag"
version = "1.0.0"
description = "Intelligent project monitor with 30-min heartbeat digests and RAG-powered knowledge base"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Hadar", email = "hadar@example.com" }]
keywords = ["rag", "llm", "productivity", "slack", "github", "notion"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# ---------------------------------------------------------------------------
# Strict dependencies
# ---------------------------------------------------------------------------
dependencies = [
    # Web framework
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    # LLM clients
    "anthropic>=0.28.0,<1.0.0",
    # Vector store
    "faiss-cpu>=1.8.0,<2.0.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    # Config & validation
    "pydantic>=2.7.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    # HTTP
    "httpx>=0.27.0,<1.0.0",
    # Resilience (retry + rate-limiting)
    "tenacity>=8.3.0,<10.0.0",
    # Scheduler
    "apscheduler>=3.10.4,<4.0.0",
    # Dashboard
    "streamlit>=1.36.0,<2.0.0",
    # YouTube
    "youtube-transcript-api>=0.6.2,<1.0.0",
    # Numerics
    "numpy>=1.26.0,<3.0.0",
]

# ---------------------------------------------------------------------------
# Optional dev/test extras
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",          # for TestClient
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# ---------------------------------------------------------------------------
# CLI entry points
# ---------------------------------------------------------------------------
[project.scripts]
pulseboard-start  = "cli:start"
pulseboard-config = "cli:setup_wizard"

# ---------------------------------------------------------------------------
# Setuptools build — flat layout (all source at project root)
# ---------------------------------------------------------------------------
[tool.setuptools]
py-modules = ["config", "models", "llm", "main", "scheduler", "dashboard", "demo_data", "cli"]

[tool.setuptools.packages.find]
where = ["."]
include = ["heartbeat*", "integrations*", "rag*", "routes*"]

# ---------------------------------------------------------------------------
# Ruff linting
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
